# backend/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SettingsResponse": {
        "properties": {
          "bm25_weight": {
            "title": "Bm25 Weight",
            "type": "number"
          },
          "chunk_overlap": {
            "title": "Chunk Overlap",
            "type": "integer"
          },
          "chunk_size": {
            "title": "Chunk Size",
            "type": "integer"
          },
          "context_window": {
            "title": "Context Window",
            "type": "integer"
          },
          "embedding_model_name": {
            "title": "Embedding Model Name",
            "type": "string"
          },
          "hybrid_search": {
            "title": "Hybrid Search",
            "type": "boolean"
          },
          "llm_api_key_configured": {
            "title": "Llm Api Key Configured",
            "type": "boolean"
          },
          "llm_base_url": {
            "title": "Llm Base Url",
            "type": "string"
          },
          "llm_model_name": {
            "title": "Llm Model Name",
            "type": "string"
          },
          "max_tokens": {
            "title": "Max Tokens",
            "type": "integer"
          },
          "query_rewrite_enabled": {
            "title": "Query Rewrite Enabled",
            "type": "boolean"
          },
          "rerank_enabled": {
            "title": "Rerank Enabled",
            "type": "boolean"
          },
          "rerank_threshold": {
            "title": "Rerank Threshold",
            "type": "number"
          },
          "rerank_top_k": {
            "title": "Rerank Top K",
            "type": "integer"
          },
          "retrieval_top_k": {
            "title": "Retrieval Top K",
            "type": "integer"
          },
          "similarity_threshold": {
            "title": "Similarity Threshold",
            "type": "number"
          },
          "temperature": {
            "title": "Temperature",
            "type": "number"
          },
          "top_k": {
            "title": "Top K",
            "type": "integer"
          },
          "web_search_enabled": {
            "title": "Web Search Enabled",
            "type": "boolean"
          }
        },
        "required": [
          "llm_base_url",
          "llm_model_name",
          "embedding_model_name",
          "chunk_size",
          "chunk_overlap",
          "top_k",
          "temperature",
          "max_tokens",
          "context_window",
          "similarity_threshold",
          "hybrid_search",
          "bm25_weight",
          "retrieval_top_k",
          "rerank_top_k",
          "rerank_enabled",
          "rerank_threshold",
          "query_rewrite_enabled",
          "web_search_enabled",
          "llm_api_key_configured"
        ],
        "title": "SettingsResponse",
        "type": "object"
      },
      "SettingsUpdate": {
        "properties": {
          "bm25_weight": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bm25 Weight"
          },
          "context_window": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context Window"
          },
          "hybrid_search": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hybrid Search"
          },
          "llm_base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Llm Base Url"
          },
          "llm_model_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Llm Model Name"
          },
          "max_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Tokens"
          },
          "query_rewrite_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Query Rewrite Enabled"
          },
          "rerank_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rerank Enabled"
          },
          "rerank_threshold": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rerank Threshold"
          },
          "rerank_top_k": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rerank Top K"
          },
          "retrieval_top_k": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retrieval Top K"
          },
          "similarity_threshold": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Similarity Threshold"
          },
          "temperature": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature"
          },
          "top_k": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Top K"
          },
          "web_search_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Web Search Enabled"
          }
        },
        "title": "SettingsUpdate",
        "type": "object"
      },
      "TagCreate": {
        "properties": {
          "color": {
            "default": "default",
            "title": "Color",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "TagCreate",
        "type": "object"
      },
...
      "TagResponse": {
        "properties": {
          "color": {
            "title": "Color",
            "type": "string"
          },
          "doc_count": {
            "default": 0,
            "title": "Doc Count",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "color"
        ],
        "title": "TagResponse",
        "type": "object"
      },
      "TagUpdate": {
        "properties": {
          "color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "title": "TagUpdate",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "LocalRAG",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/settings": {
      "get": {
        "operationId": "get_settings_api_settings_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Settings",
        "tags": [
          "settings"
        ]
      },
      "put": {
        "operationId": "update_settings_api_settings_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Settings",
        "tags": [
          "settings"
        ]
      }
    },
    "/api/tags": {
      "get": {
        "description": "列出当前用户的标签，附带文档数量。",
        "operationId": "list_tags_api_tags_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TagResponse"
                  },
                  "title": "Response List Tags Api Tags Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Tags",
        "tags": [
          "tags"
        ]
      },
      "post": {
        "description": "创建标签。",
        "operationId": "create_tag_api_tags_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TagCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Tag",
        "tags": [
          "tags"
        ]
      }
    },
    "/api/tags/attach": {
      "post": {
        "description": "为文档添加标签。",
        "operationId": "attach_tag_api_tags_attach_post",
        "parameters": [
          {
            "in": "query",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "tag_id",
            "required": true,
            "schema": {
              "title": "Tag Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Attach Tag",
        "tags": [
          "tags"
        ]
      }
    },
    "/api/tags/detach": {
      "post": {
        "description": "移除文档的标签。",
        "operationId": "detach_tag_api_tags_detach_post",
        "parameters": [
          {
            "in": "query",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "tag_id",
            "required": true,
            "schema": {
              "title": "Tag Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Detach Tag",
        "tags": [
          "tags"
        ]
      }
    },
    "/api/tags/{tag_id}": {
      "delete": {
        "description": "删除标签（同时移除所有文档关联）。",
        "operationId": "delete_tag_api_tags__tag_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "tag_id",
            "required": true,
            "schema": {
              "title": "Tag Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Tag",
        "tags": [
          "tags"
        ]
      },
      "put": {
        "description": "更新标签。",
        "operationId": "update_tag_api_tags__tag_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "tag_id",
            "required": true,
            "schema": {
              "title": "Tag Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TagUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Tag",
        "tags": [
          "tags"
        ]
      }
    }
  }
}